+++ Coordinator/mlxbar/__init__.py
@@ -1,3 +1,3 @@
 """MLXBar coordinator package."""
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"